+++ test/integration/boot-nodes.ts
@@ -1,12 +1,12 @@
 import "mocha" // using @types/mocha
 import { expect } from "chai"
-import { Network } from "../.."
 
 import { DVoteGateway, Web3Gateway } from "../../src/net/gateway"
 import {
     getGatewaysFromBootNodeData,
     fetchFromBootNode,
-    fetchDefaultBootNode
+    fetchDefaultBootNode,
+    NetworkID
 } from "../../src/net/gateway-bootnodes"
 
 const DEV_BOOTNODES_URL = "https://bootnodes.vocdoni.net/gateways.dev.json"
@@ -32,7 +32,7 @@ describe("Boot nodes", () => {
             expect(typeof bootnodes[networkId].dvote[0].disconnect).to.equal("function")
             expect(typeof bootnodes[networkId].dvote[0].getUri).to.equal("function")
             expect(typeof await bootnodes[networkId].dvote[0].getUri()).to.equal("string")
-            expect(typeof bootnodes[networkId].dvote[0].sendMessage).to.equal("function")
+            expect(typeof bootnodes[networkId].dvote[0].sendRequest).to.equal("function")
             expect(typeof bootnodes[networkId].dvote[0].isConnected).to.equal("function")
             expect(typeof bootnodes[networkId].dvote[0].publicKey).to.equal("string")
 
@@ -49,109 +49,113 @@ describe("Boot nodes", () => {
             expect(typeof networkId).to.equal("string")
         }
 
-        expect(bootnodes["xdai"].dvote).to.have.length.gte(1)
-        expect(bootnodes["xdai"].web3).to.have.length.gte(1)
+        const NETWORK_ID = "xdai"
+        expect(bootnodes[NETWORK_ID].dvote).to.have.length.gte(1)
+        expect(bootnodes[NETWORK_ID].web3).to.have.length.gte(1)
 
-        expect(bootnodes["xdai"].dvote[0] instanceof DVoteGateway).to.be.true
-        expect(bootnodes["xdai"].web3[0] instanceof Web3Gateway).to.be.true
+        expect(bootnodes[NETWORK_ID].dvote[0] instanceof DVoteGateway).to.be.true
+        expect(bootnodes[NETWORK_ID].web3[0] instanceof Web3Gateway).to.be.true
 
-        expect(typeof bootnodes["xdai"].dvote[0].connect).to.equal("function")
-        expect(typeof bootnodes["xdai"].dvote[0].disconnect).to.equal("function")
-        expect(typeof bootnodes["xdai"].dvote[0].getUri).to.equal("function")
-        expect(typeof await bootnodes["xdai"].dvote[0].getUri()).to.equal("string")
-        expect(typeof bootnodes["xdai"].dvote[0].sendMessage).to.equal("function")
-        expect(typeof bootnodes["xdai"].dvote[0].isConnected).to.equal("function")
-        expect(typeof bootnodes["xdai"].dvote[0].publicKey).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].connect).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].disconnect).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].getUri).to.equal("function")
+        expect(typeof await bootnodes[NETWORK_ID].dvote[0].getUri()).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].sendRequest).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].isConnected).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].publicKey).to.equal("string")
 
-        expect(typeof bootnodes["xdai"].web3[0].attach).to.equal("function")
-        expect(typeof bootnodes["xdai"].web3[0].deploy).to.equal("function")
-        expect(typeof bootnodes["xdai"].web3[0].getProvider).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].attach).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].deploy).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].getProvider).to.equal("function")
 
         // XDAI Stage
 
         bootnodeData = await fetchFromBootNode(STAGE_BOOTNODES_URL)
-        const options = {testing:true}
+        const options = { testing: true }
         bootnodes = await getGatewaysFromBootNodeData(bootnodeData, options)
 
         for (let networkId in bootnodes) {
             expect(typeof networkId).to.equal("string")
         }
 
-        expect(bootnodes["xdai"].dvote).to.have.length.gte(1)
-        expect(bootnodes["xdai"].web3).to.have.length.gte(1)
+        expect(bootnodes[NETWORK_ID].dvote).to.have.length.gte(1)
+        expect(bootnodes[NETWORK_ID].web3).to.have.length.gte(1)
 
-        expect(bootnodes["xdai"].dvote[0] instanceof DVoteGateway).to.be.true
-        expect(bootnodes["xdai"].web3[0] instanceof Web3Gateway).to.be.true
+        expect(bootnodes[NETWORK_ID].dvote[0] instanceof DVoteGateway).to.be.true
+        expect(bootnodes[NETWORK_ID].web3[0] instanceof Web3Gateway).to.be.true
 
-        expect(typeof bootnodes["xdai"].dvote[0].connect).to.equal("function")
-        expect(typeof bootnodes["xdai"].dvote[0].disconnect).to.equal("function")
-        expect(typeof bootnodes["xdai"].dvote[0].getUri).to.equal("function")
-        expect(typeof await bootnodes["xdai"].dvote[0].getUri()).to.equal("string")
-        expect(typeof bootnodes["xdai"].dvote[0].sendMessage).to.equal("function")
-        expect(typeof bootnodes["xdai"].dvote[0].isConnected).to.equal("function")
-        expect(typeof bootnodes["xdai"].dvote[0].publicKey).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].connect).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].disconnect).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].getUri).to.equal("function")
+        expect(typeof await bootnodes[NETWORK_ID].dvote[0].getUri()).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].sendRequest).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].isConnected).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].publicKey).to.equal("string")
 
-        expect(typeof bootnodes["xdai"].web3[0].attach).to.equal("function")
-        expect(typeof bootnodes["xdai"].web3[0].deploy).to.equal("function")
-        expect(typeof bootnodes["xdai"].web3[0].getProvider).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].attach).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].deploy).to.equal("function")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].getProvider).to.equal("function")
     }).timeout(20000)
 
     it("fetchDefaultBootNode (default) should provide a bootnode JSON structure", async () => {
-        let bootnodes = await fetchDefaultBootNode("sokol")
+        let NETWORK_ID: NetworkID = "xdai"
+        let bootnodes = await fetchDefaultBootNode(NETWORK_ID)
 
         for (let networkId in bootnodes) {
             expect(typeof networkId).to.equal("string")
         }
 
-        expect(Array.isArray(bootnodes["sokol"].dvote)).to.be.true
-        expect(typeof bootnodes["sokol"].dvote[0].uri).to.equal("string")
-        expect(typeof bootnodes["sokol"].dvote[0].pubKey).to.equal("string")
-        expect(Array.isArray(bootnodes["sokol"].dvote[0].apis)).to.be.true
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote)).to.be.true
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].pubKey).to.equal("string")
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote[0].apis)).to.be.true
 
-        expect(typeof bootnodes["sokol"].web3[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].uri).to.equal("string")
 
         // XDAI Stage
-        const options = {testing:true}
-        bootnodes = await fetchDefaultBootNode("xdai", options)
+        const options = { testing: true }
+        bootnodes = await fetchDefaultBootNode(NETWORK_ID, options)
         for (let networkId in bootnodes) {
             expect(typeof networkId).to.equal("string")
         }
 
-        expect(Array.isArray(bootnodes["xdai"].dvote)).to.be.true
-        expect(typeof bootnodes["xdai"].dvote[0].uri).to.equal("string")
-        expect(typeof bootnodes["xdai"].dvote[0].pubKey).to.equal("string")
-        expect(Array.isArray(bootnodes["xdai"].dvote[0].apis)).to.be.true
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote)).to.be.true
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].pubKey).to.equal("string")
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote[0].apis)).to.be.true
 
-        expect(typeof bootnodes["xdai"].web3[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].uri).to.equal("string")
 
         // XDAI
-        bootnodes = await fetchDefaultBootNode("xdai")
+        bootnodes = await fetchDefaultBootNode(NETWORK_ID)
         for (let networkId in bootnodes) {
             expect(typeof networkId).to.equal("string")
         }
 
-        expect(Array.isArray(bootnodes["xdai"].dvote)).to.be.true
-        expect(typeof bootnodes["xdai"].dvote[0].uri).to.equal("string")
-        expect(typeof bootnodes["xdai"].dvote[0].pubKey).to.equal("string")
-        expect(Array.isArray(bootnodes["xdai"].dvote[0].apis)).to.be.true
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote)).to.be.true
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].pubKey).to.equal("string")
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote[0].apis)).to.be.true
 
-        expect(typeof bootnodes["xdai"].web3[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].uri).to.equal("string")
 
         // SOKOL
-        bootnodes = await fetchDefaultBootNode("sokol")
+        NETWORK_ID = "sokol"
+        bootnodes = await fetchDefaultBootNode(NETWORK_ID)
         for (let networkId in bootnodes) {
             expect(typeof networkId).to.equal("string")
         }
 
-        expect(Array.isArray(bootnodes["sokol"].dvote)).to.be.true
-        expect(typeof bootnodes["sokol"].dvote[0].uri).to.equal("string")
-        expect(typeof bootnodes["sokol"].dvote[0].pubKey).to.equal("string")
-        expect(Array.isArray(bootnodes["sokol"].dvote[0].apis)).to.be.true
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote)).to.be.true
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].pubKey).to.equal("string")
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote[0].apis)).to.be.true
 
-        expect(typeof bootnodes["sokol"].web3[0].uri).to.equal("string")
-    }).timeout(30000)
+        expect(typeof bootnodes[NETWORK_ID].web3[0].uri).to.equal("string")
+    }).timeout(20000)
 
     it("fetchFromBootNode should provide a bootnode JSON structure", async () => {
+        const NETWORK_ID = "xdai"
         let bootnodes = await fetchFromBootNode(DEV_BOOTNODES_URL)
 
         for (let networkId in bootnodes) {
@@ -172,26 +176,26 @@ describe("Boot nodes", () => {
             expect(typeof networkId).to.equal("string")
         }
 
-        expect(Array.isArray(bootnodes["xdai"].dvote)).to.be.true
-        expect(typeof bootnodes["xdai"].dvote[0].uri).to.equal("string")
-        expect(typeof bootnodes["xdai"].dvote[0].pubKey).to.equal("string")
-        expect(Array.isArray(bootnodes["xdai"].dvote[0].apis)).to.be.true
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote)).to.be.true
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].pubKey).to.equal("string")
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote[0].apis)).to.be.true
 
-        expect(typeof bootnodes["xdai"].web3[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].uri).to.equal("string")
 
-         // XDAI Stage
-         bootnodes = await fetchFromBootNode(STAGE_BOOTNODES_URL)
+        // XDAI Stage
+        bootnodes = await fetchFromBootNode(STAGE_BOOTNODES_URL)
 
-         for (let networkId in bootnodes) {
-             expect(typeof networkId).to.equal("string")
-         }
+        for (let networkId in bootnodes) {
+            expect(typeof networkId).to.equal("string")
+        }
 
-         expect(Array.isArray(bootnodes["xdai"].dvote)).to.be.true
-         expect(typeof bootnodes["xdai"].dvote[0].uri).to.equal("string")
-         expect(typeof bootnodes["xdai"].dvote[0].pubKey).to.equal("string")
-         expect(Array.isArray(bootnodes["xdai"].dvote[0].apis)).to.be.true
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote)).to.be.true
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].dvote[0].pubKey).to.equal("string")
+        expect(Array.isArray(bootnodes[NETWORK_ID].dvote[0].apis)).to.be.true
 
-         expect(typeof bootnodes["xdai"].web3[0].uri).to.equal("string")
+        expect(typeof bootnodes[NETWORK_ID].web3[0].uri).to.equal("string")
     }).timeout(20000)
 
 })