+++ test/mocks/web-socket-service.ts
@@ -4,24 +4,24 @@ import { signJsonBody } from "../../src/util/json-sign"
 import { DVoteGateway } from "../../src/net/gateway"
 
 export type WSResponse = {
-    id: string
-    response: {
-        ok: boolean
-        request: string  // Request ID here as well
-        message?: string
-        timestamp?: number
-        [key: string]: any
-    }
-    signature?: string
+    id: string,
+    response: WSResponseBody,
+    signature: string
+}
+export type WSResponseBody = {
+    ok: boolean
+    message?: string
+    timestamp?: number
+    [key: string]: any
 }
 export type WebSocketMockedInteraction = {
-    actual?: any,                     // What the client actually sent
-    responseData: WSResponse     // What to send as a response
+    requested?: any,                 // What the client actually sent
+    responseData: WSResponseBody     // What to send as a response
 }
 
 const defaultPort = 8500
 const defaultMnemonic = "myth like bonus scare over problem client lizard pioneer submit female collect"
-const defaultConnectResponse: WSResponse = { id: "dummy", response: { request: "dummy", timestamp: 123, ok: true, apiList: ["file", "vote", "census", "results"], health: 100 }, signature: "" }
+const defaultConnectResponse: WSResponseBody = { timestamp: 123, ok: true, apiList: ["file", "vote", "census", "results"], health: 100 }
 
 // THE GATEWAY SERVER MOCK
 
@@ -37,10 +37,12 @@ export class DevWebSocketServer {
     public interactionCount: number = 0
     public readonly wallet: Wallet
 
-    constructor(params: { port?: number, responses?: WSResponse[] } = {}) {
+    constructor(params: { port?: number, responses?: WSResponseBody[] } = {}) {
+        if (params.responses && params.responses.some(res => typeof res != "object")) throw new Error("Invalid response provided")
+
         this.port = params.port || defaultPort
         this.interactionList = (params.responses || [defaultConnectResponse]).map(response => ({
-            actual: null,             // no requests received yet
+            requested: null,             // no requests received yet
             responseData: response
         }))
 
@@ -67,20 +69,22 @@ export class DevWebSocketServer {
         if (idx >= this.interactionList.length) throw new Error("The Gateway received more transactions than it should: " + (this.interactionCount + 1))
         else if (!this.interactionList[idx]) throw new Error("Mock transaction data is empty")
 
-        this.interactionList[idx].actual = JSON.parse(requestData)
+        this.interactionList[idx].requested = JSON.parse(requestData)
 
         if (!this.activeSocket) throw new Error("No socket client to reply to")
 
-        const responseData: WSResponse = this.interactionList[idx].responseData
-        responseData.id = this.interactionList[idx].actual.id
+        const response: WSResponseBody = this.interactionList[idx].responseData
+        response.request = this.interactionList[idx].requested.id
+        response.timestamp = Math.floor(Date.now() / 1000)
+
+        const responseData: WSResponse = {
+            id: this.interactionList[idx].requested.id,
+            response,
+            signature: ""
+        }
         responseData.signature = await signJsonBody(responseData.response, this.wallet)
 
-        if (!responseData.response) {
-            console.error("GATEWAY MOCK: No request field present", responseData)
-            throw new Error("Invalid mock response set")
-        }
-
-        responseData.response.request = this.interactionList[idx].actual.id
+        responseData.response.request = this.interactionList[idx].requested.id
 
         // Reply stringified
         this.activeSocket.send(JSON.stringify(responseData))
@@ -88,9 +92,11 @@ export class DevWebSocketServer {
         this.interactionCount++
     }
 
-    public addResponse(res: WSResponse) {
+    public addResponse(res: WSResponseBody) {
+        if (typeof res != "object") throw new Error("Invalid response provided")
+
         this.interactionList.push({
-            actual: null,      // no request received yet
+            requested: null,      // no request received yet
             responseData: res
         })
     }